+++ samplerprep/drivers/__init__.py
@@ -12,6 +12,9 @@ _DRIVER_MAP = {
     "queen-of-pentacles": "queen",
     "bitbox": "bitbox",
     "squid": "squid",
+    "assimil8or": "assimil8or",
+    "clutch": "clutch",
+    "addac112": "addac112",
 }